Resurrect QSerialPort::restoreParametersOnClose property
The property was deprecated in Qt 5.3 times, and later removed when
migrating to Qt 6. However, the internal flag, as well as the logic of
caching the previous port parameters, was still available in the code.
The flag was simply hardcoded to true, meaning that the port was always
trying to restore its settings on close.

However, there are some usecases, when the serial port parameters are
used to control the behavior of the connected device. For example, the
linked Jira ticket mentions that some custom device can be reset by
setting a specific baud rate. In such case, restoring the baud rate
before closing the port prevents the device from resetting.

This patch restores the public getters and setters for the property.
The internal logic of restoring the values was already there, so the
patch does not touch that part. However, it updates the initialization
logic, which failed to restore the original values if some of the
initialization steps failed. It also adds a unit-test to check that
the new property is handled correctly.

This patch partly reverts 36c0c8d92adffbaecb3c195e1a40eff67857e19e.

[ChangeLog][QSerialPort] Restored the
QSerialPort::settingsRestoredOnClose property.

Fixes: QTBUG-129756
Fixes: QTBUG-107440
Change-Id: Ia464d466438f5ea24f4376c17dda517a7f9bc1ab
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>