QSerialPort: document that Space/Mark parity is not always supported
Space and Mark parity is unsupported on UNIX systems where CMSPAR flag
is not supported. One example of such systems is macOS.
The old QSerialPort class implementation made some attempts to emulate
reading and writing, but they never worked properly, and so they were
removed later (at least partly, the write emulation still exists).

Let's be honest and document the current state.
Also add a warning when attempting to set unsupported parity. Add the
warning to a custom logging category, so that the users could silence
it if they absolutely want to do it.

Task-number: QTBUG-131679
Pick-to: 6.9
Change-Id: I4e22f2dd25f4b042b8c3c37921f87e0b44414b35
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>